Moving BasicMath method to test folder as a stub method
