+++ tests/php/BasicMath/MethodTest.php
@@ -3,9 +3,9 @@
 namespace SilverStripe\MFA\Tests\BasicMath;
 
 use SilverStripe\Dev\SapphireTest;
-use SilverStripe\MFA\BasicMath\Method;
 use SilverStripe\MFA\Method\Handler\LoginHandlerInterface;
 use SilverStripe\MFA\Method\Handler\RegisterHandlerInterface;
+use SilverStripe\MFA\Tests\Stub\BasicMath\Method;
 
 class MethodTest extends SapphireTest
 {
