+++ tests/php/Stub/BasicMath/MethodLoginHandler.php
@@ -1,15 +1,16 @@
 <?php
-namespace SilverStripe\MFA\BasicMath;
+namespace SilverStripe\MFA\Tests\Stub\BasicMath;
 
 use SilverStripe\Control\HTTPRequest;
 use SilverStripe\Core\Config\Configurable;
+use SilverStripe\Dev\TestOnly;
 use SilverStripe\MFA\Method\Handler\LoginHandlerInterface;
 use SilverStripe\MFA\Store\StoreInterface;
 
 /**
  * Handles login attempts for the Math Method
  */
-class MethodLoginHandler implements LoginHandlerInterface
+class MethodLoginHandler implements LoginHandlerInterface, TestOnly
 {
     use Configurable;
